elegant-chatbot/requirements.txt
# Core dependencies
numpy>=1.21.0
pyaudio>=0.2.11
openai>=0.27.0
faster-whisper>=0.10.0  # Much faster than openai-whisper
pyttsx3>=2.90  # Fallback TTS
SpeechRecognition>=3.8.0  # Fallback STT

# Optional but recommended
# git+https://github.com/resemble-ai/chatterbox.git  # Better TTS

# Development dependencies
pytest>=7.0.0
black>=22.0.0
pylint>=2.15.0